+++ dubbo-config/dubbo-config-api/src/main/java/org/apache/dubbo/config/bootstrap/DubboBootstrap.java
@@ -186,6 +186,8 @@ public final class DubboBootstrap {
         });
         // register DubboBootstrap bean
         applicationModel.getBeanFactory().registerBean(this);
+
+        DubboShutdownHook.getInstance().register();
     }
 
     private void notifyStarted(ApplicationModel applicationModel) {
